modified resources and tests

+++ app/models.py
@@ -31,6 +31,7 @@ tables:
             student_id (int, primary_key)
 """
 import re
+
 from .application import app, db
 from sqlalchemy.exc import IntegrityError, DataError
 
@@ -39,7 +40,13 @@ def is_group_name_fits(name):
     return re.search("[a-z][a-z]-[0-9][0-9]", name)
 
 
-class Mixin:
+students_courses_relation = db.Table('students_courses_relation',
+                        db.Column('course_id', db.Integer, db.ForeignKey('courses.id'), primary_key=True),
+                        db.Column('student_id', db.Integer, db.ForeignKey('students.id'), primary_key=True)
+                        )
+
+
+class DatabaseFunctionsMixin(object):
 
     @classmethod
     def get_item(cls, item_id):
@@ -48,14 +55,34 @@ class Mixin:
     @classmethod
     def delete_item(cls, item_id):
         item = cls.query.filter_by(id=item_id).first()
+        if not item:
+            return f'item {item_id} was not found in {cls.__tablename__} table.'
         db.session.delete(item)
         db.session.commit()
-        return 'deleted item {}.'.format(item_id)
+
+        return f'deleted item {item_id}.'
 
     @classmethod
     def get_all_items_params_dict(cls):
         items = cls.query.all()
         return [item.get_params_dict() for item in items]
+
+
+    @classmethod
+    def post_item(cls, **params):
+        columns_name_list = cls.__table__.columns.keys()
+        columns_name_list.remove('id')
+        for column_name in columns_name_list:
+            assert column_name in params, '`{}` parameter missed'.format(column_name)
+
+        new_item = cls(**{column_name: params[column_name] for column_name in columns_name_list})
+
+        try:
+            db.session.add(new_item)
+            db.session.commit()
+        except IntegrityError:
+            db.session.rollback()
+            raise AssertionError('incorrect data.')
 
     def put_params(self, **params):
         for column_name in self.__table__.columns.keys():
@@ -72,24 +99,31 @@ class Mixin:
         return {column_name: getattr(self, column_name)
                 for column_name in self.__table__.columns.keys()}
 
-    @classmethod
-    def post_item(cls, **params):
-        columns_name_list = cls.__table__.columns.keys()
-        columns_name_list.remove('id')
-        for column_name in columns_name_list:
-            assert column_name in params, '{} parameter missed'.format(column_name)
 
-        new_item = cls(**{column_name: params[column_name] for column_name in columns_name_list})
+class StudentModel(db.Model, DatabaseFunctionsMixin):
+    __tablename__ = 'students'
 
-        try:
-            db.session.add(new_item)
-            db.session.commit()
-        except IntegrityError:
-            db.session.rollback()
-            raise AssertionError('incorrect data.')
+    id = db.Column(db.Integer, primary_key=True)
+    group_id = db.Column(db.Integer, db.ForeignKey('groups.id'))
+    first_name = db.Column(db.String)
+    last_name = db.Column(db.String)
+    courses = db.relationship('CourseModel', secondary=students_courses_relation, lazy='subquery',
+                              backref=db.backref('students', lazy=True))
+
+    def __init__(self, group_id, first_name, last_name):
+        self.group_id = group_id
+        self.first_name = first_name
+        self.last_name = last_name
+
+    def get_params_dict(self):
+        params_dict = super(StudentModel, self).get_params_dict()
+
+        courses = db.session.query(students_courses_relation).filter_by(student_id=self.id)
+        params_dict['courses_ids'] = [course.course_id for course in courses]
+        return params_dict
 
 
-class GroupModel(db.Model, Mixin):
+class GroupModel(db.Model, DatabaseFunctionsMixin):
     __tablename__ = 'groups'
 
     id = db.Column(db.Integer, primary_key=True)
@@ -119,7 +153,7 @@ class GroupModel(db.Model, Mixin):
             'cannot delete the group with students. Student ids: {}' \
                 .format(', '.join([str(student.id) for student in students]))
 
-        super(GroupModel, cls).delete_item(item_id)
+        return super(GroupModel, cls).delete_item(item_id)
 
     @classmethod
     def post_item(cls, **params):
@@ -128,36 +162,7 @@ class GroupModel(db.Model, Mixin):
         super(GroupModel, cls).post_item(**params)
 
 
-students_courses_relation = db.Table('students_courses_relation',
-                        db.Column('course_id', db.Integer, db.ForeignKey('courses.id'), primary_key=True),
-                        db.Column('student_id', db.Integer, db.ForeignKey('students.id'), primary_key=True)
-                        )
-
-
-class StudentModel(db.Model, Mixin):
-    __tablename__ = 'students'
-
-    id = db.Column(db.Integer, primary_key=True)
-    group_id = db.Column(db.Integer, db.ForeignKey('groups.id'))
-    first_name = db.Column(db.String)
-    last_name = db.Column(db.String)
-    courses = db.relationship('CourseModel', secondary=students_courses_relation, lazy='subquery',
-                              backref=db.backref('students', lazy=True))
-
-    def __init__(self, group_id, first_name, last_name):
-        self.group_id = group_id
-        self.first_name = first_name
-        self.last_name = last_name
-
-    def get_params_dict(self):
-        params_dict = super(StudentModel, self).get_params_dict()
-
-        courses = db.session.query(students_courses_relation).filter_by(student_id=self.id)
-        params_dict['courses_ids'] = [course.course_id for course in courses]
-        return params_dict
-
-
-class CourseModel(db.Model, Mixin):
+class CourseModel(db.Model, DatabaseFunctionsMixin):
     __tablename__ = 'courses'
 
     id = db.Column(db.Integer, primary_key=True)
